Include `Paths_biscuit_haskell` to `autogen-modules`
As instructed by hackage:

```
Error: Invalid package

Packages using 'cabal-version: 2.0' and the autogenerated module Paths_* must
include it also on the 'autogen-modules' field besides 'exposed-modules' and
'other-modules'. This specifies that the module does not come with the package
and is generated on setup. Modules built with a custom Setup.hs script also go
here to ensure that commands like sdist don't fail.
```