Provide a type-safe way to separate the sig check from the datalog verif
Before, it was possible to just check the signature without verifying the datalog,
but the function returned a boolean, which meant users had to ensure that the
signature was correctly checked by their own means.

This commit gives a type-safe way to separate sig check from datalog verification with
an opaque `ValidBiscuit` type.