v2: introduce a scoped executor for Datalog
According to the biscuit v2 _feuille de route_, the datalog is now
evaluated in a staged process: each block is ran in sequence, with
its checks. This way, a new block cannot generate new facts that
could interfere with previous blocks rules and checks.

In effects, this makes `#authority` and `#ambient` flagging
automatic, and generalized to every block, not just the authority.