biscuit(v2): make biscuit functions polymorphic on proof type
`validateBiscuit` and friends don't care at all about the proof type
since it has already been validated.

`serializeBiscuit` and friends already know how to handle both open
and sealed tokens, it was just a matter of automatically projecting
Open and Sealed tokens to the main case, with a little typeclass.

This allows to remove most of the `fromOpen` / `fromSealed` uses.