biscuit: expose the data that matched allow or deny queries
Instead of just returning the query itself, the error (for
deny policies) and success (for allow policies) data types now also
carry the bindings that matched the query. This will make reading
info from a token easier.

For instance, consider a macaroon-like setup where the token carries
a user id, and some checks:

```
user("1234");
check if operation("read");
check if resource("file1");
```

The authorizer can be `allow if user($user);` instead of `allow if true;`:
this will make sure that the token carries a user id that can be used by the
handler.