make datalog evaluation fail on malformed rule
Previously, malformed rules (rules where the head had unbound variables) were
silently ignored. Now it makes evaluation fail completely, which is a safer
default (and consistent with the rust implementation).

While a rule being silently ignored is okay with `checks` and `allow` policies,
it can be dangerous if it was intended to be used with a `deny` policy