3rd-party: add external signatures to PB & Crypto
For now the high-level API requires a full biscuit to compute both the
regular and external signatures. In the future, it would be useful to
have a helper with weaker requirements, that would only take what's
needed so that the 3rd party can provide the payload & signature without
getting access to the whole biscuit.