3rd-party: decouple authority scope from authorizer scope
Previously, authority and authorizer where in the same scope. That made sense
with first-party blocks, where we always trust the authority in any context.
With third-party blocks however, that gave rise to subtle corner cases:
facts from the current block are always in scope: a block always trusts itself.
That meant, hovewer, that in the authorizer, the authority block was always trusted,
even when specifying `trusting {pk};` (note the absence of the `authority` keyword).

With this change, authority and authorizer are now separated: the authorizer lives in a special
block, appended after the other blocks.

Both the current block and the authorizer are always trusted. The rationale is that:

- a block always trusts itself
- the whole computation happens under the control of the authorizer, so it does not make
  sense to not trust it

The default behaviour stays the same (when no `trusting` annotation is added): the current block,
the authorizer and the authority are trusted.