biscuit: allow parameter substitution for sets
The existing `ToTerm` typeclass asked implementations to work in all contexts
that made it convenient from a user point of view, but restricted its use to
values that could appear anywhere. It could thus not be used to produce a set
(which can only happen at the top level). The solution was to make the allowed
contexts part of the typeclass parameters. Types that can go anywhere can let
the `PredicateOrFact` and `IsWithinSet` parameters open, while types that have
specific constraints (like sets) can fix the contexts in which they're allowed.

The call site then decides which instances are allowed based on its actual
context instead of asking too much.