biscuit: rewrite the Datalog parser
Closes https://github.com/biscuit-auth/biscuit-haskell/issues/62

The main goal is to provide better error messages by:

- using megaparsec instead of attoparsec
- parse syntactically valid elements and then reject invalid semantics,
  instead of directly refusing to parse syntactically correct but
  semantically incorrect elements

As a result, the parser structure is quite simplified. Instead of polymorphic,
context-dependent parsers configured via typeclasses, parsers are monomorphic,
with explicit configuration by arguments when needed.

Since this is a complete rewrite, it also fixes previous issues / incompatibilities :

- operator precedences are consistent with https://github.com/biscuit-auth/biscuit/issues/116
- ed25519 public key literals are consistent with https://github.com/biscuit-auth/biscuit/pull/117

An additional feature (not stable yet, may change soon) is runtime parsing of datalog elements, with
support for parameters (previously it was only possible at compile-time).