biscuit-servant: relax types for error handlers
The error handlers forced a polymorphic return type. In practice, that meant only
being able to return a `ServerError` through `throwError`.

For `handleBiscuitWith`, this just makes the signature simpler
without actually changing what's possible: `handleBiscuitWith` is
made to be used as a natural transformation, so the return type is
still polymorphic.

For `checkBiscuitWith` and `checkBiscuitMWith`, which are made to
be used in handlers directly, this will allow recovering from within
the error handler.